+++ examples/complete-alb/main.tf
@@ -27,7 +27,7 @@ data "aws_route53_zone" "this" {
 
 module "security_group" {
   source  = "terraform-aws-modules/security-group/aws"
-  version = "~> 3.0"
+  version = "~> 4.0"
 
   name        = "alb-sg-${random_pet.this.id}"
   description = "Security group for example usage with ALB"
@@ -50,7 +50,7 @@ module "security_group" {
 
 module "acm" {
   source  = "terraform-aws-modules/acm/aws"
-  version = "~> 2.0"
+  version = "~> 3.0"
 
   domain_name = local.domain_name # trimsuffix(data.aws_route53_zone.this.name, ".") # Terraform >= 0.12.17
   zone_id     = data.aws_route53_zone.this.id
@@ -89,12 +89,12 @@ module "alb" {
   load_balancer_type = "application"
 
   vpc_id          = data.aws_vpc.default.id
-  security_groups = [module.security_group.this_security_group_id]
+  security_groups = [module.security_group.security_group_id]
   subnets         = data.aws_subnet_ids.all.ids
 
   #   # See notes in README (ref: https://github.com/terraform-providers/terraform-provider-aws/issues/7987)
   #   access_logs = {
-  #     bucket = module.log_bucket.this_s3_bucket_id
+  #     bucket = module.log_bucket.s3_bucket_id
   #   }
 
   http_tcp_listeners = [
@@ -131,7 +131,7 @@ module "alb" {
     {
       port               = 443
       protocol           = "HTTPS"
-      certificate_arn    = module.acm.this_acm_certificate_arn
+      certificate_arn    = module.acm.acm_certificate_arn
       target_group_index = 1
     },
     # Authentication actions only allowed with HTTPS
@@ -140,7 +140,7 @@ module "alb" {
       protocol           = "HTTPS"
       action_type        = "authenticate-cognito"
       target_group_index = 1
-      certificate_arn    = module.acm.this_acm_certificate_arn
+      certificate_arn    = module.acm.acm_certificate_arn
       authenticate_cognito = {
         authentication_request_extra_params = {
           display = "page"
@@ -159,7 +159,7 @@ module "alb" {
       protocol           = "HTTPS"
       action_type        = "authenticate-oidc"
       target_group_index = 1
-      certificate_arn    = module.acm.this_acm_certificate_arn
+      certificate_arn    = module.acm.acm_certificate_arn
       authenticate_oidc = {
         authentication_request_extra_params = {
           display = "page"
@@ -248,7 +248,75 @@ module "alb" {
     },
     {
       https_listener_index = 0
+      priority             = 4
+
+      actions = [{
+        type = "weighted-forward"
+        target_groups = [
+          {
+            target_group_index = 1
+            weight             = 2
+          },
+          {
+            target_group_index = 0
+            weight             = 1
+          }
+        ]
+        stickiness = {
+          enabled  = true
+          duration = 3600
+        }
+      }]
+
+      conditions = [{
+        query_strings = [{
+          key   = "weighted"
+          value = "true"
+        }]
+      }]
+    },
+    {
+      https_listener_index = 0
       priority             = 5000
+      actions = [{
+        type        = "redirect"
+        status_code = "HTTP_302"
+        host        = "www.youtube.com"
+        path        = "/watch"
+        query       = "v=dQw4w9WgXcQ"
+        protocol    = "HTTPS"
+      }]
+
+      conditions = [{
+        query_strings = [{
+          key   = "video"
+          value = "random"
+        }]
+      }]
+    },
+  ]
+
+  http_tcp_listener_rules = [
+    {
+      http_tcp_listener_index = 0
+      priority                = 3
+      actions = [{
+        type         = "fixed-response"
+        content_type = "text/plain"
+        status_code  = 200
+        message_body = "This is a fixed response"
+      }]
+
+      conditions = [{
+        http_headers = [{
+          http_header_name = "x-Gimme-Fixed-Response"
+          values           = ["yes", "please", "right now"]
+        }]
+      }]
+    },
+    {
+      http_tcp_listener_index = 0
+      priority                = 5000
       actions = [{
         type        = "redirect"
         status_code = "HTTP_302"
@@ -285,6 +353,17 @@ module "alb" {
         protocol            = "HTTP"
         matcher             = "200-399"
       }
+      protocol_version = "HTTP1"
+      targets = {
+        my_ec2 = {
+          target_id = aws_instance.this.id
+          port      = 80
+        },
+        my_ec2_again = {
+          target_id = aws_instance.this.id
+          port      = 8080
+        }
+      }
       tags = {
         InstanceTargetGroupTag = "baz"
       }
@@ -293,6 +372,16 @@ module "alb" {
       name_prefix                        = "l1-"
       target_type                        = "lambda"
       lambda_multi_value_headers_enabled = true
+      targets = {
+        # Lambda function permission should be granted before
+        # it is used. There can be an error:
+        # NB: Error registering targets with target group:
+        # AccessDenied: elasticloadbalancing principal does not
+        # have permission to invoke ... from target group ...
+        my_lambda = {
+          target_id = module.lambda_function.lambda_function_arn
+        }
+      }
     },
   ]
 
@@ -307,6 +396,18 @@ module "alb" {
   target_group_tags = {
     MyGlobalTargetGroupTag = "bar"
   }
+
+  https_listener_rules_tags = {
+    MyLoadBalancerHTTPSListenerRule = "bar"
+  }
+
+  https_listeners_tags = {
+    MyLoadBalancerHTTPSListener = "bar"
+  }
+
+  http_tcp_listeners_tags = {
+    MyLoadBalancerTCPListener = "bar"
+  }
 }
 
 #########################
@@ -316,4 +417,77 @@ module "lb_disabled" {
   source = "../../"
 
   create_lb = false
+}
+
+##################
+# Extra resources
+##################
+data "aws_ami" "amazon_linux" {
+  most_recent = true
+
+  owners = ["amazon"]
+
+  filter {
+    name = "name"
+
+    values = [
+      "amzn-ami-hvm-*-x86_64-gp2",
+    ]
+  }
+
+  filter {
+    name = "owner-alias"
+
+    values = [
+      "amazon",
+    ]
+  }
+}
+
+resource "aws_instance" "this" {
+  ami           = data.aws_ami.amazon_linux.id
+  instance_type = "t3.nano"
+}
+
+#############################################
+# Using packaged function from Lambda module
+#############################################
+
+locals {
+  package_url = "https://raw.githubusercontent.com/terraform-aws-modules/terraform-aws-lambda/master/examples/fixtures/python3.8-zip/existing_package.zip"
+  downloaded  = "downloaded_package_${md5(local.package_url)}.zip"
+}
+
+resource "null_resource" "download_package" {
+  triggers = {
+    downloaded = local.downloaded
+  }
+
+  provisioner "local-exec" {
+    command = "curl -L -o ${local.downloaded} ${local.package_url}"
+  }
+}
+
+module "lambda_function" {
+  source  = "terraform-aws-modules/lambda/aws"
+  version = "~> 2.0"
+
+  function_name = "${random_pet.this.id}-lambda"
+  description   = "My awesome lambda function"
+  handler       = "index.lambda_handler"
+  runtime       = "python3.8"
+
+  publish = true
+
+  create_package         = false
+  local_existing_package = local.downloaded
+
+  allowed_triggers = {
+    AllowExecutionFromELB = {
+      service    = "elasticloadbalancing"
+      source_arn = module.alb.target_group_arns[1] # index should match the correct target_group
+    }
+  }
+
+  depends_on = [null_resource.download_package]
 }