+++ main.tf
@@ -31,8 +31,10 @@ resource "aws_lb" "this" {
     for_each = var.subnet_mapping
 
     content {
-      subnet_id     = subnet_mapping.value.subnet_id
-      allocation_id = lookup(subnet_mapping.value, "allocation_id", null)
+      subnet_id            = subnet_mapping.value.subnet_id
+      allocation_id        = lookup(subnet_mapping.value, "allocation_id", null)
+      private_ipv4_address = lookup(subnet_mapping.value, "private_ipv4_address", null)
+      ipv6_address         = lookup(subnet_mapping.value, "ipv6_address", null)
     }
   }
 
@@ -57,16 +59,18 @@ resource "aws_lb_target_group" "main" {
   name        = lookup(var.target_groups[count.index], "name", null)
   name_prefix = lookup(var.target_groups[count.index], "name_prefix", null)
 
-  vpc_id      = var.vpc_id
-  port        = lookup(var.target_groups[count.index], "backend_port", null)
-  protocol    = lookup(var.target_groups[count.index], "backend_protocol", null) != null ? upper(lookup(var.target_groups[count.index], "backend_protocol")) : null
-  target_type = lookup(var.target_groups[count.index], "target_type", null)
+  vpc_id           = var.vpc_id
+  port             = lookup(var.target_groups[count.index], "backend_port", null)
+  protocol         = lookup(var.target_groups[count.index], "backend_protocol", null) != null ? upper(lookup(var.target_groups[count.index], "backend_protocol")) : null
+  protocol_version = lookup(var.target_groups[count.index], "protocol_version", null) != null ? upper(lookup(var.target_groups[count.index], "protocol_version")) : null
+  target_type      = lookup(var.target_groups[count.index], "target_type", null)
 
   deregistration_delay               = lookup(var.target_groups[count.index], "deregistration_delay", null)
   slow_start                         = lookup(var.target_groups[count.index], "slow_start", null)
   proxy_protocol_v2                  = lookup(var.target_groups[count.index], "proxy_protocol_v2", false)
   lambda_multi_value_headers_enabled = lookup(var.target_groups[count.index], "lambda_multi_value_headers_enabled", false)
   load_balancing_algorithm_type      = lookup(var.target_groups[count.index], "load_balancing_algorithm_type", null)
+  preserve_client_ip                 = lookup(var.target_groups[count.index], "preserve_client_ip", null)
 
   dynamic "health_check" {
     for_each = length(keys(lookup(var.target_groups[count.index], "health_check", {}))) == 0 ? [] : [lookup(var.target_groups[count.index], "health_check", {})]
@@ -103,11 +107,35 @@ resource "aws_lb_target_group" "main" {
     },
   )
 
-  depends_on = [aws_lb.this]
-
   lifecycle {
     create_before_destroy = true
   }
+}
+
+locals {
+  # Merge the target group index into a product map of the targets so we
+  # can figure out what target group we should attach each target to.
+  # Target indexes can be dynamically defined, but need to match
+  # the function argument reference. This means any additional arguments
+  # can be added later and only need to be updated in the attachment resource below.
+  # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/lb_target_group_attachment#argument-reference
+  target_group_attachments = merge(flatten([
+    for index, group in var.target_groups : [
+      for k, targets in group : {
+        for target_key, target in targets : join(".", [index, target_key]) => merge({ tg_index = index }, target)
+      }
+      if k == "targets"
+    ]
+  ])...)
+}
+
+resource "aws_lb_target_group_attachment" "this" {
+  for_each = var.create_lb && local.target_group_attachments != null ? local.target_group_attachments : {}
+
+  target_group_arn  = aws_lb_target_group.main[each.value.tg_index].arn
+  target_id         = each.value.target_id
+  port              = lookup(each.value, "port", null)
+  availability_zone = lookup(each.value, "availability_zone", null)
 }
 
 resource "aws_lb_listener_rule" "https_listener_rule" {
@@ -219,6 +247,37 @@ resource "aws_lb_listener_rule" "https_listener_rule" {
     }
   }
 
+  # weighted forward actions
+  dynamic "action" {
+    for_each = [
+      for action_rule in var.https_listener_rules[count.index].actions :
+      action_rule
+      if action_rule.type == "weighted-forward"
+    ]
+
+    content {
+      type = "forward"
+      forward {
+        dynamic "target_group" {
+          for_each = action.value["target_groups"]
+
+          content {
+            arn    = aws_lb_target_group.main[target_group.value["target_group_index"]].id
+            weight = target_group.value["weight"]
+          }
+        }
+        dynamic "stickiness" {
+          for_each = [lookup(action.value, "stickiness", {})]
+
+          content {
+            enabled  = try(stickiness.value["enabled"], false)
+            duration = try(stickiness.value["duration"], 1)
+          }
+        }
+      }
+    }
+  }
+
   # Path Pattern condition
   dynamic "condition" {
     for_each = [
@@ -318,8 +377,179 @@ resource "aws_lb_listener_rule" "https_listener_rule" {
       }
     }
   }
+
+  tags = merge(
+    var.tags,
+    var.https_listener_rules_tags,
+    lookup(var.https_listener_rules[count.index], "tags", {}),
+  )
 }
 
+resource "aws_lb_listener_rule" "http_tcp_listener_rule" {
+  count = var.create_lb ? length(var.http_tcp_listener_rules) : 0
+
+  listener_arn = aws_lb_listener.frontend_http_tcp[lookup(var.http_tcp_listener_rules[count.index], "http_tcp_listener_index", count.index)].arn
+  priority     = lookup(var.http_tcp_listener_rules[count.index], "priority", null)
+
+  # redirect actions
+  dynamic "action" {
+    for_each = [
+      for action_rule in var.http_tcp_listener_rules[count.index].actions :
+      action_rule
+      if action_rule.type == "redirect"
+    ]
+
+    content {
+      type = action.value["type"]
+      redirect {
+        host        = lookup(action.value, "host", null)
+        path        = lookup(action.value, "path", null)
+        port        = lookup(action.value, "port", null)
+        protocol    = lookup(action.value, "protocol", null)
+        query       = lookup(action.value, "query", null)
+        status_code = action.value["status_code"]
+      }
+    }
+  }
+
+  # fixed-response actions
+  dynamic "action" {
+    for_each = [
+      for action_rule in var.http_tcp_listener_rules[count.index].actions :
+      action_rule
+      if action_rule.type == "fixed-response"
+    ]
+
+    content {
+      type = action.value["type"]
+      fixed_response {
+        message_body = lookup(action.value, "message_body", null)
+        status_code  = lookup(action.value, "status_code", null)
+        content_type = action.value["content_type"]
+      }
+    }
+  }
+
+  # forward actions
+  dynamic "action" {
+    for_each = [
+      for action_rule in var.http_tcp_listener_rules[count.index].actions :
+      action_rule
+      if action_rule.type == "forward"
+    ]
+
+    content {
+      type             = action.value["type"]
+      target_group_arn = aws_lb_target_group.main[lookup(action.value, "target_group_index", count.index)].id
+    }
+  }
+
+  # Path Pattern condition
+  dynamic "condition" {
+    for_each = [
+      for condition_rule in var.http_tcp_listener_rules[count.index].conditions :
+      condition_rule
+      if length(lookup(condition_rule, "path_patterns", [])) > 0
+    ]
+
+    content {
+      path_pattern {
+        values = condition.value["path_patterns"]
+      }
+    }
+  }
+
+  # Host header condition
+  dynamic "condition" {
+    for_each = [
+      for condition_rule in var.http_tcp_listener_rules[count.index].conditions :
+      condition_rule
+      if length(lookup(condition_rule, "host_headers", [])) > 0
+    ]
+
+    content {
+      host_header {
+        values = condition.value["host_headers"]
+      }
+    }
+  }
+
+  # Http header condition
+  dynamic "condition" {
+    for_each = [
+      for condition_rule in var.http_tcp_listener_rules[count.index].conditions :
+      condition_rule
+      if length(lookup(condition_rule, "http_headers", [])) > 0
+    ]
+
+    content {
+      dynamic "http_header" {
+        for_each = condition.value["http_headers"]
+
+        content {
+          http_header_name = http_header.value["http_header_name"]
+          values           = http_header.value["values"]
+        }
+      }
+    }
+  }
+
+  # Http request method condition
+  dynamic "condition" {
+    for_each = [
+      for condition_rule in var.http_tcp_listener_rules[count.index].conditions :
+      condition_rule
+      if length(lookup(condition_rule, "http_request_methods", [])) > 0
+    ]
+
+    content {
+      http_request_method {
+        values = condition.value["http_request_methods"]
+      }
+    }
+  }
+
+  # Query string condition
+  dynamic "condition" {
+    for_each = [
+      for condition_rule in var.http_tcp_listener_rules[count.index].conditions :
+      condition_rule
+      if length(lookup(condition_rule, "query_strings", [])) > 0
+    ]
+
+    content {
+      dynamic "query_string" {
+        for_each = condition.value["query_strings"]
+
+        content {
+          key   = lookup(query_string.value, "key", null)
+          value = query_string.value["value"]
+        }
+      }
+    }
+  }
+
+  # Source IP address condition
+  dynamic "condition" {
+    for_each = [
+      for condition_rule in var.http_tcp_listener_rules[count.index].conditions :
+      condition_rule
+      if length(lookup(condition_rule, "source_ips", [])) > 0
+    ]
+
+    content {
+      source_ip {
+        values = condition.value["source_ips"]
+      }
+    }
+  }
+
+  tags = merge(
+    var.tags,
+    var.http_tcp_listener_rules_tags,
+    lookup(var.http_tcp_listener_rules[count.index], "tags", {}),
+  )
+}
 
 resource "aws_lb_listener" "frontend_http_tcp" {
   count = var.create_lb ? length(var.http_tcp_listeners) : 0
@@ -361,6 +591,12 @@ resource "aws_lb_listener" "frontend_http_tcp" {
       }
     }
   }
+
+  tags = merge(
+    var.tags,
+    var.http_tcp_listeners_tags,
+    lookup(var.http_tcp_listeners[count.index], "tags", {}),
+  )
 }
 
 resource "aws_lb_listener" "frontend_https" {
@@ -372,6 +608,7 @@ resource "aws_lb_listener" "frontend_https" {
   protocol        = lookup(var.https_listeners[count.index], "protocol", "HTTPS")
   certificate_arn = var.https_listeners[count.index]["certificate_arn"]
   ssl_policy      = lookup(var.https_listeners[count.index], "ssl_policy", var.listener_ssl_policy_default)
+  alpn_policy     = lookup(var.https_listeners[count.index], "alpn_policy", null)
 
   dynamic "default_action" {
     for_each = length(keys(var.https_listeners[count.index])) == 0 ? [] : [var.https_listeners[count.index]]
@@ -449,6 +686,12 @@ resource "aws_lb_listener" "frontend_https" {
       target_group_arn = aws_lb_target_group.main[lookup(default_action.value, "target_group_index", count.index)].id
     }
   }
+
+  tags = merge(
+    var.tags,
+    var.https_listeners_tags,
+    lookup(var.https_listeners[count.index], "tags", {}),
+  )
 }
 
 resource "aws_lb_listener_certificate" "https_listener" {