+++ outputs.tf
@@ -1,24 +1,24 @@
-output "this_lb_id" {
+output "lb_id" {
   description = "The ID and ARN of the load balancer we created."
   value       = concat(aws_lb.this.*.id, [""])[0]
 }
 
-output "this_lb_arn" {
+output "lb_arn" {
   description = "The ID and ARN of the load balancer we created."
   value       = concat(aws_lb.this.*.arn, [""])[0]
 }
 
-output "this_lb_dns_name" {
+output "lb_dns_name" {
   description = "The DNS name of the load balancer."
   value       = concat(aws_lb.this.*.dns_name, [""])[0]
 }
 
-output "this_lb_arn_suffix" {
+output "lb_arn_suffix" {
   description = "ARN suffix of our load balancer - can be used with CloudWatch."
   value       = concat(aws_lb.this.*.arn_suffix, [""])[0]
 }
 
-output "this_lb_zone_id" {
+output "lb_zone_id" {
   description = "The zone_id of the load balancer to assist with creating DNS records."
   value       = concat(aws_lb.this.*.zone_id, [""])[0]
 }
@@ -56,4 +56,11 @@ output "target_group_arn_suffixes" {
 output "target_group_names" {
   description = "Name of the target group. Useful for passing to your CodeDeploy Deployment Group."
   value       = aws_lb_target_group.main.*.name
+}
+
+output "target_group_attachments" {
+  description = "ARNs of the target group attachment IDs."
+  value = {
+    for k, v in aws_lb_target_group_attachment.this : k => v.id
+  }
 }